==> R/00_doit.R <==
# Run all scripts
# ------------------------------------------------------------------------------
source(file = "R/01_load.R")
source(file = "R/02_clean.R")
source(file = "R/03_augment.R")
source(file = "R/04_analysis_aim.R")
source(file = "R/04_analysis_exploratory_figures.R")
source(file = "R/04_analysis_flowcharts.R")
source(file = "R/04_analysis_global_maps.R")
source(file = "R/04_analysis_linear_regression.R")
source(file = "R/04_analysis_PCA.R")
source(file = "R/04_analysis_point_plots.R")
source(file = "R/04_analysis_survival.R")
